Add files via upload
First, run VideoChat.py from ide then VideoChat2.py from a different terminal. then enter your target IP to the given text boxes of windows, then click on start listening from the first window of VideoChat.py and click start camera stream of VideoChat2.py and done. I have tested it on a single machine and fail to test because only one camera is available so we can run the first program on one computer and the second on another.